perf: remove DashboardMain function , and transfer all dashbobard Components to dashboard
