+++ resources/js/pages/admin/dashboard/CoworkingRequests.jsx
@@ -9,7 +9,7 @@ const CoworkingRequests = () => {
         <div className="rounded-xl bg-white p-5 shadow-md">
             <div className="mb-4 flex items-center justify-between">
                 <h2 className="text-xl font-bold flex items-center gap-x-5"><Briefcase color='#ffc803' />Coworking Requests</h2>
-                <Link href="/admin/coworking" className="rounded bg-beta px-3 py-1 text-white hover:bg-alpha hover:text-black">
+                <Link href="/admin/coworking" className="rounded-lg bg-beta px-3 py-1 text-white hover:bg-alpha hover:text-black">
                     View All
                 </Link>{' '}
             </div>
@@ -19,7 +19,7 @@ const CoworkingRequests = () => {
             ) : (
                 <div className="overflow-x-auto">
                     <table className="w-full table-auto border-collapse text-sm">
-                        <thead className="bg-gray-200">
+                        <thead className="bg-gray-200 ">
                             <tr>
                                 <th className="px-4 py-3 text-left font-medium text-black">Name</th>
                                 <th className="px-4 py-3 text-left font-medium text-black">Email</th>
@@ -48,7 +48,7 @@ const CoworkingRequests = () => {
                                     <td className="px-4 py-3 text-gray-600">{dayjs(request.created_at).format('YYYY-MM-DD')}</td>
                                     <td className="px-4 py-3 text-gray-600">
                                         <Link
-                                            className="rounded bg-beta px-5 py-2 text-white hover:bg-alpha hover:text-black"
+                                            className="rounded-lg bg-beta px-5 py-2 text-white hover:bg-alpha hover:text-black"
                                             href={'/admin/coworking/' + request.id}
                                         >
                                             Details
