+++ resources/js/pages/admin/dashboard/UnreadMessages.jsx
@@ -18,7 +18,7 @@ const UnreadMessages = () => {
                     </svg>
                     <h2 className="text-xl font-bold">Recent Messages</h2>
                 </div>
-                <Link href="/admin/contacts" className="rounded bg-beta px-3 py-1 text-white hover:text-black hover:bg-alpha">
+                <Link href="/admin/contacts" className="rounded-lg bg-beta px-3 py-1 text-white hover:text-black hover:bg-alpha">
                     View All
                 </Link>
             </div>
