+++ resources/js/pages/admin/dashboard/charts/Charts.jsx
