+++ resources/js/pages/admin/dashboard/charts/components/BarChart.tsx
